Reduce resource wasting in init_timeout test
Since this test is configured to fail after a very short `init_timeout`,
NimblePool tries to restart the service a bunch of times, each time
sending a `createBrowserContext` message to Chrome.

This patch increases the timeout to 10ms, so at least we don't create a
hundred of them, but maybe 10 or so.